+++ src/Enum.js
@@ -9,7 +9,6 @@ export var timeout = 100
 export var network = timeout * 6
 
 // element identifiers
-export var fallback = 0
 export var iterable = 1
 export var component = 2
 export var target = 3