+++ src/Render.js
@@ -38,7 +38,7 @@ export function dispatch (element, target, callback) {
  */
 export function resolve (fiber, element, target, value) {
 	if (target === value) {
-		target[Enum.identifier] = Node.create(fiber, element, element, element, null)
+		Interface.register(Node.create(fiber, element, element, element, null), target)
 	} else {
 		Reconcile.children(fiber, element, element, element.children, value, 0)
 	}